base.sh: fix waitForUserLogout for F42
There seems to be a change in behaviour from Fedora 41 to Fedora 42.
When a script is executed remotely a non-interactive session is created.
However, the user session with `noTTY` in Fedora F42 is picked up by the
`who`, `w` & `users` commands.

Since we run a remote script that checks for logged in users with the
`who` command, the `waitForUserLogout` function blocks and creates an
endless loop. This happens because the session only closes once the
remote script has finished executing.

The fix is a bit ugly, but we can filter this session out by checking
for user sessions that don't have a terminal device `?`