base.sh: fix loop math
I noticed this error in one of the CICD runs:

base.sh: line 38: ((: > 0 : syntax error: operand expected (error token is "> 0 ")

This is caused probably by the last change in this function which added
grep in the chain of the commands. But when grep does not match
anything, 1 is returned and nothing is on the standard output. That then
cannot be used in the shell expression in its naked form. This fixes it.

I noticed the same thing in the other function, but that one uses
"pgrep" which always prints a number. But just for consistency, I am
changing it there as well so we prevent copy-pasting this unsafe style.